refactor: worker files and folder structure refactored
