+++ worker/tsconfig.tsbuildinfo
@@ -1,13 +1,1 @@
-{
-  "root": [
-    "./src/index.ts",
-    "./src/config/index.ts",
-    "./src/services/docs.service.ts",
-    "./src/services/mail.service.ts",
-    "./src/services/model.service.ts",
-    "./src/services/scraper.service.ts",
-    "./src/services/sheets.service.ts",
-    "./src/utils/parser.ts"
-  ],
-  "version": "5.7.3"
-}
+{"root":["./src/index.ts","./src/server.ts","./src/config/index.ts","./src/processors/template.processor.ts","./src/processors/workflow.processor.ts","./src/queue/redis.queue.ts","./src/services/docs.service.ts","./src/services/mail.service.ts","./src/services/model.service.ts","./src/services/scraper.service.ts","./src/services/sheets.service.ts","./src/utils/parser.ts"],"version":"5.7.3"}
